Split TDD Steps into modules
