+++ compliance-test-kit/BDD-and-TDD/testData/productcatalog-static-roles-v1beta3/values.yaml
@@ -30,3 +30,6 @@ partyrole:
   image: lesterthomas/partyroleapi:1.0
   initializationimage: lesterthomas/roleinitialization:0.3
   versionlabel: partyroleapi-1.0
+metrics:
+  image: lesterthomas/registerallevents:0.2
+  versionlabel: lesterthomas-metricsapi-0.2
